% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_polynomial_trend_surface}
\alias{calc_polynomial_trend_surface}
\title{fit a higher-order GLM to our spatial data and a field of your choice
and use it to generate a polynomial trend raster surface of that field}
\usage{
calc_polynomial_trend_surface(pts, order = 3, field = NULL,
  predRaster = NULL)
}
\description{
fit a higher-order GLM to our spatial data and a field of your choice
and use it to generate a polynomial trend raster surface of that field
}
