% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_saturated_thickness}
\alias{calc_saturated_thickness}
\title{calculate saturated thickness for a series of well points and the base elevation of
the aquifer (as returned by ogallala::generateBaseElevationRaster())}
\usage{
calc_saturated_thickness(wellPts = NULL, baseRaster = NULL,
  surfaceRaster = NULL, convert_to_imperial = T)
}
\description{
calculate saturated thickness for a series of well points and the base elevation of
the aquifer (as returned by ogallala::generateBaseElevationRaster())
}
