% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculateSaturatedThickness}
\alias{calculateSaturatedThickness}
\title{calculate saturated thickness for a series of well points and the base elevation of
the aquifer (as returned by ogallala::generateBaseElevationRaster())}
\usage{
calculateSaturatedThickness(wellPts = NULL, baseRaster = NULL)
}
\description{
calculate saturated thickness for a series of well points and the base elevation of
the aquifer (as returned by ogallala::generateBaseElevationRaster())
}
