% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{downsample_aquifer_boundary}
\alias{downsample_aquifer_boundary}
\title{downsample well points that occur along the aquifer boundary}
\usage{
downsample_aquifer_boundary(wellPoints = NULL, boundary = NULL,
  width = 3000)
}
\description{
downsample well points that occur along the aquifer boundary
}
