% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{downsampleAlongAquiferBoundary}
\alias{downsampleAlongAquiferBoundary}
\title{downsample well points that occur along the aquifer boundary}
\usage{
downsampleAlongAquiferBoundary(wellPoints = NULL, boundary = NULL,
  width = 3000)
}
\description{
downsample well points that occur along the aquifer boundary
}
