% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generate_aquifer_boundary_buffer}
\alias{generate_aquifer_boundary_buffer}
\title{generate a uniform buffer region around the HP aquifer boundary
that can be used for downsampling}
\usage{
generate_aquifer_boundary_buffer(boundary = NULL, width = 5000)
}
\description{
generate a uniform buffer region around the HP aquifer boundary
that can be used for downsampling
}
