% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generate_target_raster_grid}
\alias{generate_target_raster_grid}
\title{generate a universal grid at the target resolution (500m) originally specified
by V. McGuire and others for their interpolation.}
\usage{
generate_target_raster_grid(s = NULL)
}
\description{
generate a universal grid at the target resolution (500m) originally specified
by V. McGuire and others for their interpolation.
}
