% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generate_zero_burnin_surface}
\alias{generate_zero_burnin_surface}
\title{Do a burn-in of an input raster and the ofr99-266 dry areas
dataset}
\usage{
generate_zero_burnin_surface(r = NULL, width = 500)
}
\description{
Do a burn-in of an input raster and the ofr99-266 dry areas
dataset
}
