% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generateAquiferBoundaryBuffer}
\alias{generateAquiferBoundaryBuffer}
\title{generate a uniform buffer region around the HP aquifer boundary
that can be used for downsampling}
\usage{
generateAquiferBoundaryBuffer(boundary = NULL, width = 5000)
}
\description{
generate a uniform buffer region around the HP aquifer boundary
that can be used for downsampling
}
