% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generateBaseElevationRaster}
\alias{generateBaseElevationRaster}
\title{using ofr98-393, make a contour line to raster product. The units on elevation are feet (imperial).}
\usage{
generateBaseElevationRaster(s = NULL, targetRasterGrid = NULL, mask = F)
}
\description{
using ofr98-393, make a contour line to raster product. The units on elevation are feet (imperial).
}
