% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generatePseudoZeros}
\alias{generatePseudoZeros}
\title{download the ofr99-266 dry areas dataset and randomly generate points
within the polygon features to use as pseudo-zero sat. thickness data}
\usage{
generatePseudoZeros(wellPts = NULL, targetRasterGrid = NULL, size = NULL)
}
\description{
download the ofr99-266 dry areas dataset and randomly generate points
within the polygon features to use as pseudo-zero sat. thickness data
}
