% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generateTargetRasterGrid}
\alias{generateTargetRasterGrid}
\title{generate a universal grid at the target resolution (500m) originally specified
by V. McGuire and others for their interpolation.}
\usage{
generateTargetRasterGrid(s = NULL)
}
\description{
generate a universal grid at the target resolution (500m) originally specified
by V. McGuire and others for their interpolation.
}
