% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{generateZeroBurnInSurface}
\alias{generateZeroBurnInSurface}
\title{Do a burn-in of an input raster and the ofr99-266 dry areas
dataset}
\usage{
generateZeroBurnInSurface(r = NULL, width = 500)
}
\description{
Do a burn-in of an input raster and the ofr99-266 dry areas
dataset
}
