% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{idw_interpolator}
\alias{idw_interpolator}
\title{hidden shortcut for gstat's idw interpolator}
\usage{
idw_interpolator(pts, targetRasterGrid = NULL, field = "ELEV")
}
\description{
hidden shortcut for gstat's idw interpolator
}
