% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{knn_point_smoother}
\alias{knn_point_smoother}
\title{use a KNN classifier fit to lat/lon to select a neighborhood of points around
each well and calculates a summary statistic of your choosing (e.g., mean)}
\usage{
knn_point_smoother(pts = NULL, field = NULL, k = 4, fun = mean)
}
\description{
use a KNN classifier fit to lat/lon to select a neighborhood of points around
each well and calculates a summary statistic of your choosing (e.g., mean)
}
