% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{m_glm_spatial_trend}
\alias{m_glm_spatial_trend}
\title{testing for a standard GLM with polynomial terms on latitude
and longitude}
\usage{
m_glm_spatial_trend(pts, order = 2, field = NULL)
}
\description{
testing for a standard GLM with polynomial terms on latitude
and longitude
}
