% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{m_knn_weights}
\alias{m_knn_weights}
\title{testing for a spatially-weighted GLM that attempts to down-weight
clustered records and up-weight diffuse records using KNN.}
\usage{
m_knn_weights(pts, order = 4, field = NULL, k = 5)
}
\description{
testing for a spatially-weighted GLM that attempts to down-weight
clustered records and up-weight diffuse records using KNN.
}
