% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{minMaxNormalize}
\alias{minMaxNormalize}
\title{normalize a vector X so that it's values occur on a 0-to-1 scale. Optionally
project the trend to a new maximum using the to= argument. This is useful
for rescaling predicted thickness to the known ranges of saturated thickness
throughout the HP region [0 to about 1,200 ft] (Weeks and Gutentag, 1981)}
\usage{
minMaxNormalize(x, to = 1300)
}
\description{
normalize a vector X so that it's values occur on a 0-to-1 scale. Optionally
project the trend to a new maximum using the to= argument. This is useful
for rescaling predicted thickness to the known ranges of saturated thickness
throughout the HP region [0 to about 1,200 ft] (Weeks and Gutentag, 1981)
}
