% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{mKNNWeights}
\alias{mKNNWeights}
\title{testing for a spatially-weighted GLM that attempts to down-weight
clustered records and up-weight diffuse records using KNN.}
\usage{
mKNNWeights(pts, order = 4, field = NULL, k = 5)
}
\description{
testing for a spatially-weighted GLM that attempts to down-weight
clustered records and up-weight diffuse records using KNN.
}
