% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{mSpatialTrend}
\alias{mSpatialTrend}
\title{testing for a standard GLM with polynomial terms on latitude
and longitude}
\usage{
mSpatialTrend(pts, order = 4, field = NULL)
}
\description{
testing for a standard GLM with polynomial terms on latitude
and longitude
}
