% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{parse_hrefs}
\alias{parse_hrefs}
\title{grab all zips by default and parse through what the user requested}
\usage{
parse_hrefs(url = NULL)
}
\description{
grab all zips by default and parse through what the user requested
}
