% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{parse_url}
\alias{parse_url}
\title{hidden function that grabs the full strings within a set of parantheses
in an HTML href.}
\usage{
parse_url(x = NULL, pattern = NULL)
}
\arguments{
\item{pattern}{optional pattern to filter href links against}
}
\description{
hidden function that grabs the full strings within a set of parantheses
in an HTML href.
}
