% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotResiduals}
\alias{plotResiduals}
\title{plot the z-standardized residual error for a model run}
\usage{
plotResiduals(pts, field = "residls")
}
\description{
plot the z-standardized residual error for a model run
}
