% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{polynomialTrendSurface}
\alias{polynomialTrendSurface}
\title{fit a higher-order GLM to our spatial data and a field of your choice
and use it to generate a polynomial trend raster surface of that field}
\usage{
polynomialTrendSurface(pts, order = 4, field = NULL, predRaster = NULL)
}
\description{
fit a higher-order GLM to our spatial data and a field of your choice
and use it to generate a polynomial trend raster surface of that field
}
