% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{normalize a vector x to quantiles. By default (if no quantiles are specified
for calculating a range), we will use 1 SD and produce a Z-score equivalent}
\usage{
quantile_normalize(x, quantiles = NULL, to = NULL)
}
\description{
normalize a vector x to quantiles. By default (if no quantiles are specified
for calculating a range), we will use 1 SD and produce a Z-score equivalent
}
