% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{scrape_ned}
\alias{scrape_ned}
\title{scrape the national elevation dataset (using FedData) to fetch and merge
surface elevation DEMs for our region}
\usage{
scrape_ned(s = NULL)
}
\description{
scrape the national elevation dataset (using FedData) to fetch and merge
surface elevation DEMs for our region
}
