% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{scrape_unsampled_zero_values}
\alias{scrape_unsampled_zero_values}
\title{scrape polygon features for areas reported to contain little-or-no saturated
thickness that lack well data and are thus not captured in the HP Water
Monitoring study.}
\usage{

  scrape_unsampled_zero_values(base_url = "https://water.usgs.gov/GIS/metadata/usgswrd/XML/ofr99-266.xml")
}
\description{
scrape polygon features for areas reported to contain little-or-no saturated
thickness that lack well data and are thus not captured in the HP Water
Monitoring study.
}
