% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{scrape_well_point_data}
\alias{scrape_well_point_data}
\title{scrape well-point water data from the USGS portal}
\usage{

  scrape_well_point_data(base_url = "https://ne.water.usgs.gov/projects/HPA/",
  years = 1995:(as.numeric(format(Sys.time(), "\%Y"))))
}
\description{
scrape well-point water data from the USGS portal
}
