% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{scrapeBedrockElevation}
\alias{scrapeBedrockElevation}
\title{scrape bedrock elevation data}
\usage{

  scrapeBedrockElevation(base_url = "https://water.usgs.gov/GIS/metadata/usgswrd/XML/ofr98-393_aqbase.xml")
}
\description{
scrape bedrock elevation data
}
