% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{scrapeHighPlainsAquiferBoundary}
\alias{scrapeHighPlainsAquiferBoundary}
\title{webscrape functionality for the hp aquifer boundary contour provided by USGS.
Returns a full pathname to the zipfile containing the boundary (ESRI Shapefile).}
\usage{

  scrapeHighPlainsAquiferBoundary(base_url = "https://water.usgs.gov/GIS/metadata/usgswrd/XML/ds543.xml")
}
\description{
webscrape functionality for the hp aquifer boundary contour provided by USGS.
Returns a full pathname to the zipfile containing the boundary (ESRI Shapefile).
}
