% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{scrapeNed}
\alias{scrapeNed}
\title{scrape the national elevation dataset (using FedData) to fetch and merge
surface elevation DEMs for our region}
\usage{
scrapeNed(s = NULL)
}
\description{
scrape the national elevation dataset (using FedData) to fetch and merge
surface elevation DEMs for our region
}
