% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{scrapeWellPointData}
\alias{scrapeWellPointData}
\title{scrape well-point water data from the USGS portal}
\usage{
scrapeWellPointData(base_url = "https://ne.water.usgs.gov/projects/HPA/",
  years = 1995:(as.numeric(format(Sys.time(), "\%Y"))))
}
\description{
scrape well-point water data from the USGS portal
}
