% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{split_training_testing_datasets}
\alias{split_training_testing_datasets}
\title{split an input dataset into training/testing using a user-specified ratio}
\usage{
split_training_testing_datasets(pts = NULL, split = 0.2)
}
\description{
split an input dataset into training/testing using a user-specified ratio
}
