% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{splitToTrainingTestingDatasets}
\alias{splitToTrainingTestingDatasets}
\title{split an input dataset into training/testing using a user-specified ratio}
\usage{
splitToTrainingTestingDatasets(pts = NULL, split = 0.2)
}
\description{
split an input dataset into training/testing using a user-specified ratio
}
