% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{unpack_high_plains_aquifer_boundary}
\alias{unpack_high_plains_aquifer_boundary}
\title{unpack aquifer boundary data from a zipped ESRI shapefile, as returned by
scrapeHighPlainsAquiferBoundary()}
\usage{
unpack_high_plains_aquifer_boundary(x = "ds543.zip")
}
\description{
unpack aquifer boundary data from a zipped ESRI shapefile, as returned by
scrapeHighPlainsAquiferBoundary()
}
