% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{unpack_unsampled_zero_values}
\alias{unpack_unsampled_zero_values}
\title{unpack a ofr99-266 vector dataset}
\usage{
unpack_unsampled_zero_values(x = "ofr99-266.e00")
}
\description{
unpack a ofr99-266 vector dataset
}
