% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{unpack_well_point_data}
\alias{unpack_well_point_data}
\title{hidden function that unpacks and processes well point data, returning to
user as CSV}
\usage{
unpack_well_point_data(x = NULL)
}
\description{
hidden function that unpacks and processes well point data, returning to
user as CSV
}
