% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{unpackHighPlainsAquiferBoundary}
\alias{unpackHighPlainsAquiferBoundary}
\title{unpack aquifer boundary data from a zipped ESRI shapefile, as returned by
scrapeHighPlainsAquiferBoundary()}
\usage{
unpackHighPlainsAquiferBoundary(x = "ds543.zip")
}
\description{
unpack aquifer boundary data from a zipped ESRI shapefile, as returned by
scrapeHighPlainsAquiferBoundary()
}
