% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{unpackUnsampledZeroValues}
\alias{unpackUnsampledZeroValues}
\title{unpack a ofr99-266 vector dataset}
\usage{
unpackUnsampledZeroValues(x = "ofr99-266.e00")
}
\description{
unpack a ofr99-266 vector dataset
}
