% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{unpackWellPointData}
\alias{unpackWellPointData}
\title{hidden function that unpacks and processes well point data, returning to
user as CSV}
\usage{
unpackWellPointData(x = NULL)
}
\description{
hidden function that unpacks and processes well point data, returning to
user as CSV
}
