% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{unpackWellPointData}
\alias{unpackWellPointData}
\title{hidden unpack and process well point data, returning to user as CSV}
\usage{
unpackWellPointData(x = NULL)
}
\description{
hidden unpack and process well point data, returning to user as CSV
}
